+++ src/bignum/bignum_rand_get.c
@@ -6,7 +6,7 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/22 16:59:14 by acazuc            #+#    #+#             */
-/*   Updated: 2018/07/22 19:00:27 by acazuc           ###   ########.fr       */
+/*   Updated: 2018/08/04 21:05:46 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 #include "bignum.h"
 #include "sha1.h"
 
-uint64_t	g_bignum_rand;
+uint32_t	g_bignum_rand;
 
-uint64_t	bignum_rand_get()
+uint32_t	bignum_rand_get()
 {
 	t_sha1_ctx	ctx;
 	uint8_t		digest[20];
@@ -26,5 +26,5 @@ uint64_t	bignum_rand_get()
 	sha1_init(&ctx);
 	sha1_update(&ctx, (uint8_t*)&g_bignum_rand, sizeof(g_bignum_rand));
 	sha1_final(digest, &ctx);
-	return (*(uint64_t*)digest);
+	return (*(uint32_t*)digest);
 }
