+++ src/bignum/bignum_rshift.c
@@ -6,13 +6,13 @@
 /*   By: acazuc <acazuc@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/06 14:11:52 by acazuc            #+#    #+#             */
-/*   Updated: 2018/07/09 23:42:22 by acazuc           ###   ########.fr       */
+/*   Updated: 2018/08/04 21:04:18 by acazuc           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "bignum.h"
 
-int	bignum_rshift(t_bignum *r, t_bignum *a, uint64_t n)
+int	bignum_rshift(t_bignum *r, t_bignum *a, uint32_t n)
 {
 	int	sign;
 	int	ret;
